# docs/api/openapi.yaml
openapi: 3.0.3
info:
  title: lte-system API v1
  version: 2.2.0
  description: |
    无状态 LTE 实验小区工具的标准 REST API。
    旧版零散路由（`/start`、`/stop`…）为兼容性冻结保留；新集成必须使用 `/api/v1`。
    完整参考与迁移表见 docs/API.md。
    Standard REST API for the stateless LTE lab-cell toolkit.
    Legacy ad-hoc routes (`/start`, `/stop`, …) remain frozen for
    compatibility; new integrations must use `/api/v1`.
    See docs/API.md for the full reference and migration table.
    Docker 默认监听 0.0.0.0:8081，客户端使用服务器 IP；建议设置 LTE_API_TOKEN。
    Docker listens on 0.0.0.0:8081 by default; clients use the server IP.
    A strong random LTE_API_TOKEN is recommended; LTE_LISTEN remains configurable.
    JSON request bodies contain exactly one object (not null/array/trailing data).
    Authentication failures and recovered panics are included in request-ID audit logs.
servers:
  - url: http://127.0.0.1:8081
    description: 服务器本地调用示例 / server-local request example
security:
  - bearerAuth: []
components:
  securitySchemes:
    bearerAuth:
      type: http
      scheme: bearer
      description: 仅当服务端设置 LTE_API_TOKEN 时需要 / Required only when LTE_API_TOKEN is set on the server.
  schemas:
    Envelope:
      type: object
      required: [code, message, request_id]
      properties:
        code:
          type: integer
          description: 0 表示成功，否则为 HTTP 状态码*100+序号 / 0 = success; HTTP*100+seq otherwise
          example: 0
        message:
          type: string
          example: cell started
        data:
          description: 端点负载（对象、数组或 null） / endpoint payload (object, array or null)
        request_id:
          type: string
          example: 432678718a07d689
    FieldError:
      type: object
      required: [field, reason]
      properties:
        field: { type: string, example: band }
        reason: { type: string, example: must be one of 1,3,5,7,8,34,39,40,41 }
    CellStart:
      type: object
      additionalProperties: false
      properties:
        band: { type: string, example: "7", description: "取值之一 / one of 1,3,5,7,8,34,39,40,41" }
        apn: { type: string, example: srsapn }
        mcc: { type: string, example: "001", description: "3 位数字 / 3 digits" }
        mnc: { type: string, example: "01", description: "2-3 位数字 / 2-3 digits" }
        network: { type: string, example: auto, default: auto, description: "上行策略；省略/空值/auto 按当前网络命名空间最低 metric 的可用 IPv4 默认路由解析，也可显式指定接口 / Uplink policy; omitted/empty/auto resolves the usable lowest-metric IPv4 default route in the current network namespace, or name an interface explicitly" }
        sdr: { type: string, enum: [auto, uhd, bladerf, zmq], default: auto }
        device_args: { type: string, description: "射频驱动参数，未设置时针对 B210 自动调优 / RF driver args, auto-tuned for B210 when unset" }
        tx_gain: { type: integer, minimum: 0, maximum: 90, default: 80 }
        rx_gain: { type: integer, minimum: 0, maximum: 90, default: 40 }
        n_prb: { type: integer, enum: [6, 15, 25, 50, 75, 100], default: 25 }
        full_net_name: { type: string, description: "UE 上显示的运营商名称（NITZ），1-32 字符 / operator name on UE (NITZ), 1-32 chars" }
        short_net_name: { type: string, description: "运营商短名称，1-32 字符 / short operator name, 1-32 chars" }
        dns: { type: string, description: "经 PCO 下发给 UE 的 DNS（IPv4） / DNS handed to UEs via PCO (IPv4)" }
      description: 单一 JSON 对象，上限 1 MiB；空对象 {} 复用上次成功启动的存档 / Exactly one JSON object, maximum 1 MiB; empty object {} reuses the saved profile from the last successful start.
    UE:
      type: object
      properties:
        apn: { type: string, nullable: true }
        imsi: { type: string, nullable: true }
        ip: { type: string, nullable: true }
    SIMProgram:
      type: object
      required: [imsi]
      description: 单一 JSON 对象，上限 64 KiB；未知字段忽略；已有 IMSI 的认证冲突在硬件操作前拒绝 / Exactly one JSON object, maximum 64 KiB; unknown fields are ignored; authentication conflicts for an existing IMSI are rejected before hardware work.
      properties:
        imsi: { type: string, example: "001010123456789", description: "15 位数字 / 15 digits" }
        mcc: { type: string }
        mnc: { type: string }
        mnc3: { type: boolean }
        iccid: { type: string, description: '"auto" 保留出厂值 / "auto" keeps the factory value' }
        ki: { type: string, description: "32 位十六进制 / 32 hex" }
        op: { type: string, description: "32 位十六进制，与 opc 互斥 / 32 hex, mutually exclusive with opc" }
        opc: { type: string, description: "32 位十六进制 / 32 hex" }
        op_type: { type: string, enum: [op, opc] }
        auth: { type: string, enum: [mil, xor] }
        amf: { type: string, description: "4 位十六进制 / 4 hex" }
        acc: { type: string, description: "4 位十六进制 / 4 hex" }
        adm: { type: string }
        spn: { type: string }
        name: { type: string }
        sqn: { type: string, description: "12 位十六进制 / 12 hex" }
        qci: { type: integer }
        card: { type: string }
    UEEntry:
      type: object
      properties:
        name: { type: string }
        auth: { type: string }
        imsi: { type: string }
    ConnectivityDiagnostics:
      type: object
      required: [cell, log_window, registration, pdn, nas_visibility, chap, user_plane, dns, network, limitations]
      description: 只读、有界、聚合的分层诊断；不探测外网、不停止小区、不创建抓包、不返回 IMSI 或 CHAP 凭据 / Read-only bounded aggregate layer diagnostics; does not probe the Internet, stop the cell, create captures, or return IMSI/CHAP credentials.
      properties:
        cell:
          type: object
          description: 进程状态及请求/解析后的上行接口策略 / Process state and requested/resolved uplink policy
        log_window:
          type: object
          description: EPC 日志尾窗元数据，最多读取 2 MiB / EPC log-tail metadata, maximum 2 MiB read
        registration:
          type: object
          description: 聚合 Attach/鉴权/Security Mode/Attach Complete 计数；不跨 UE 拼接 / Aggregate attach/authentication/security/attach-complete counts; never joined across UEs
        pdn:
          type: object
          description: 聚合 PDN 请求、IP 分配、承载激活、拒绝计数 / Aggregate PDN request, IP allocation, bearer activation and reject counts
        nas_visibility:
          type: object
          description: 日志窗口中的 NAS 加密可见性计数 / NAS ciphering visibility counts from the log window
        chap:
          type: object
          description: CHAP 可观察性；S1AP pcap 最大 64 MiB、tshark 3 秒且输出有界；不返回握手内容 / CHAP observability; 64 MiB S1AP cap, 3-second bounded tshark output; no handshake material returned
        user_plane:
          type: object
          description: 聚合 SGi 上下行 IP 与 DNS 请求/响应计数；不返回包内容或端点 / Aggregate SGi uplink/downlink IP and DNS request/response counts; no packet contents or endpoints returned
        dns:
          type: object
          description: DNS 配置证据与抓包观察必须分开解释 / DNS configuration and capture observations are reported separately
        network:
          type: object
          description: 当前网络命名空间内默认路由、SGi、IPv4 转发及规则的 configuration_only 证据 / configuration_only evidence for route, SGi, IPv4 forwarding and rules in the current network namespace
        limitations:
          type: array
          items: { type: string }
paths:
  /api/v1/cell:
    post:
      summary: 启动小区（空对象复用存档） / Start the cell (empty object reuses saved profile)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/CellStart' }
      responses:
        '200': { description: 启动成功（含幂等复用） / started (also on idempotent reuse) }
        '400': { description: 请求体损坏 / malformed body }
        '413': { description: JSON 请求体超过 1 MiB / JSON body exceeds 1 MiB }
        '409': { description: 小区运行中 / cell already running }
        '422': { description: 校验失败（含 data.errors） / validation failed (data.errors) }
        '503': { description: 无 SDR 附着 / no SDR attached }
    get:
      summary: 小区状态 / Cell status
      responses:
        '200': { description: '小区状态 / {running,epc,enb,pcap,started_at?,band?,apn?,net_name?,network?,resolved_network?}' }
    delete:
      summary: 停止小区并清理本服务管理的孤立抓包（幂等） / Stop cell and clean up owned orphan captures (idempotent)
      responses:
        '200': { description: '停止结果 / {stopped:true|false}' }
        '500': { description: 本服务管理的进程仍未停止 / an owned process remains running }
  /api/v1/ue:
    get:
      summary: 附着终端快照（仅第一台） / Attached UE snapshot (first UE only)
      responses:
        '200': { description: 终端信息 / UE }
        '404': { description: 无附着终端 / no UE attached }
        '412': { description: 小区未运行 / cell not running }
  /api/v1/diagnostics/connectivity:
    get:
      summary: 只读分层连通性证据 / Read-only layered connectivity evidence
      description: 返回独立的注册、PDN/IP、CHAP、SGi/DNS、路由/转发/NAT证据。部分证据缺失仍返回 200，并在对应 state/reason/problems/limitations 中说明。/ Returns independent registration, PDN/IP, CHAP, SGi/DNS and route/forwarding/NAT evidence. Missing evidence still returns 200 and is explained by component state/reason/problems/limitations.
      responses:
        '200':
          description: 有界诊断快照（允许部分来源不可用） / Bounded diagnostic snapshot (partial sources allowed)
          content:
            application/json:
              schema:
                allOf:
                  - $ref: '#/components/schemas/Envelope'
                  - type: object
                    properties:
                      data: { $ref: '#/components/schemas/ConnectivityDiagnostics' }
        '429': { description: 已有诊断运行，data.reason=diagnostic_busy / another diagnostic is running; data.reason=diagnostic_busy }
  /api/v1/crack/jobs:
    post:
      summary: 开始爆破（会停止运行中小区） / Start APN password cracking (STOPS the running cell)
      responses:
        '202': { description: 任务运行中，轮询 /api/v1/crack/result / job running, poll /api/v1/crack/result }
        '404': { description: 无终端数据 / no UE data }
        '409': { description: 任务已在运行 / job already running }
        '412': { description: 当次未采集到可用 S1AP/CHAP 前置证据 / current session has no usable S1AP/CHAP prerequisite evidence }
        '422': { description: 抓包过大、不完整或不可解码 / capture oversized, incomplete or undecodable }
        '503': { description: tshark 缺失或超时 / tshark missing or timed out }
  /api/v1/crack/result:
    get:
      summary: 爆破结果 / Cracking result
      responses:
        '200': { description: 状态 ready|running / state ready|running }
        '404': { description: 字典无口令 / 无终端数据 / password not in dictionary / no UE data }
        '412': { description: 当次未采集到可用 S1AP/CHAP 前置证据 / current session has no usable S1AP/CHAP prerequisite evidence }
        '422': { description: 抓包过大、不完整或不可解码 / capture oversized, incomplete or undecodable }
        '503': { description: tshark 缺失或超时 / tshark missing or timed out }
  /api/v1/config/subscribers:
    post:
      summary: 原子上传 user_db.csv，与写卡事务串行（重启后生效） / Atomically upload user_db.csv, serialized with SIM transactions (takes effect after restart)
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                userdb: { type: string, format: binary }
      responses:
        '200': { description: 上传成功 / uploaded }
        '413': { description: 超过 8 MiB / exceeds 8 MiB }
        '422': { description: 缺少文件 / missing file }
  /api/v1/config/wordlist:
    post:
      summary: 原子上传爆破字典 / Atomically upload crack dictionary
      requestBody:
        required: true
        content:
          multipart/form-data:
            schema:
              type: object
              properties:
                wordlist: { type: string, format: binary }
      responses:
        '200': { description: 上传成功 / uploaded }
        '413': { description: 超过 8 MiB / exceeds 8 MiB }
        '422': { description: 缺少文件 / missing file }
  /api/v1/captures/{id}:
    get:
      summary: 下载抓包 / Download a capture
      parameters:
        - in: path
          name: id
          required: true
          schema: { type: string, enum: [lte-data, s1ap, enb, epc, "0", "1", "2", "3"] }
      responses:
        '200': { description: 文件下载 / file download }
        '404': { description: 未知 id 或未就绪 / unknown id or not ready }
  /api/v1/simcards:
    post:
      summary: 串行事务写卡（冲突预检，最长约 180s） / Program a SIM in a serialized transaction with conflict preflight (up to ~180s)
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: '#/components/schemas/SIMProgram' }
      responses:
        '200': { description: 写卡成功 / programmed }
        '400': { description: 非单一 JSON 对象 / body is not exactly one JSON object }
        '413': { description: JSON 请求体超过 64 KiB / JSON body exceeds 64 KiB }
        '409': { description: 卡已在库 / card already in database }
        '412': { description: 未插卡 / no card inserted }
        '422': { description: 校验失败 / validation failed }
        '503': { description: 无读卡器 / no reader attached }
  /api/v1/profile:
    get:
      summary: 存档启动配置与终端列表（无密钥） / Saved launch profile + UE list (no key material)
      responses:
        '200': { description: '存档与终端列表 / {has_profile,profile?,ues[]}' }
  /api/v1/health:
    get:
      summary: 健康与 SDR 检测 / Health + SDR detection
      responses:
        '200': { description: '健康状态 / {ok,running,sdr,time}' }